requirements.txt
rawpy
pickle
matplotlib
numpy
kornia
scipy
opencv-python
pillow
imageio